モデルの項目間に改行を入れる等

+++ furniture-design-rest-api/src/main/java/com/furniture_design/furniture_design_rest_api/services/FurnitureServiceImlWithoutRepo.java
@@ -3,14 +3,14 @@ package com.furniture_design.furniture_design_rest_api.services;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.stereotype.Service;
 import com.furniture_design.furniture_design_rest_api.errors.NotFoundException;
 import com.furniture_design.furniture_design_rest_api.models.FurnitureItem;
 
 @Service
 public class FurnitureServiceImlWithoutRepo implements FurnitureService {
-  private final AtomicInteger _counter = new AtomicInteger();
+  private final AtomicLong _counter = new AtomicLong();
   private final List<FurnitureItem> _furnitureItems = new ArrayList<>() {
     {
       add(new FurnitureItem(_counter.incrementAndGet(), "", "furniture 1", 0, new String[0], 0, 0,
